activate AI on weeklynet
There is a new "force enable" parameter in proto alpha but it can not be
used on weeklynet because the parameter is not available in oxford,
which is weeklynet's starting proto.

An alternative is to enable the vote, set the threshold to 5% and have
the teztnets baker vote yes. This way, AI will be active ~ 8 cycles
after chain activation.

This is how we were doing for oxford 1, so the threshold is already set
to 5% and the baker is already configured to vote yes.